feat: Add Home page
