+++ db.json
@@ -7,12 +7,6 @@
       "reminder": true
     },
     {
-      "id": "2",
-      "text": "Hang out with Jonas",
-      "day": "2022/10/18 at 19:00",
-      "reminder": false
-    },
-    {
       "id": "3",
       "text": "Watch movie with le Chris",
       "day": "2022/10/21 at 18:45",
