+++ src/App.vue
@@ -5,13 +5,7 @@
         title="Task Tracker"
         :showAddTask="showAddTask"
         />
-      <div v-show="showAddTask">
-        <AddTask @add-task="addTask"/>
-      </div>
-      <Tasks 
-        @delete-task="deleteTask" 
-        @toggle-reminder="toggleReminder"
-        :tasks='tasks'/>
+      <router-view :showAddTask="showAddTask"></router-view>
       <Footer />
    </div>
 </template>
@@ -19,87 +13,23 @@
 <script>
 import Header from './components/Header';
 import Footer from './components/Footer';
-import Tasks from './components/Tasks';
-import AddTask from './components/AddTask';
 
 export default {
   name: 'App',
   components: {
     Header,
-    Footer,
-    Tasks,
-    AddTask,    
+    Footer,   
   },
   data() {
     return {
-      tasks: [],
       showAddTask: false,
     }
   },
   methods: {
-    async deleteTask(id) {
-      if (confirm('Are you sure?')) {
-        const res = await fetch(`http://localhost:5000/tasks/${id}`, {
-          method: 'DELETE',
-        })
-
-        res.status === 200 
-          ? (this.tasks = this.tasks.filter((t) => t.id !== id))
-          : alert('Error deleting task')     
-      }
-    }, 
-    async toggleReminder(id) {
-      const taskToToggle = await this.fetchTask(id)
-      const updTask = {...taskToToggle, reminder: !taskToToggle.reminder}
-
-      const res = await fetch(`http://localhost:5000/tasks/${id}`, {
-        method: 'PUT',
-        headers: {
-          'Content-type': 'application/json'
-        },
-        body: JSON.stringify(updTask)
-      })
-
-      const data = await res.json()
-
-      this.tasks = this.tasks.map((t) => 
-        t.id === id ? {...t, reminder: data.reminder} : t
-      )
-    },
-    async addTask(task) {
-      const res = await fetch('http://localhost:5000/tasks', {
-        method: 'POST',
-        headers: {
-          'Content-type': 'application/json'          
-        }, 
-        body: JSON.stringify(task)
-      })
-
-      const data = await res.json()
-      
-      this.tasks = [...this.tasks, data]
-    },
     toggleAddTask() {
       this.showAddTask = !this.showAddTask
     },
-    async fetchTasks() {
-      const res = await fetch('http://localhost:5000/tasks')
-
-      const data = await res.json()
-
-      return data
-    },
-    async fetchTask(id) {
-      const res = await fetch(`http://localhost:5000/tasks/${id}`)
-
-      const data = await res.json()
-
-      return data
-    }
-  },
-  async created() {
-    this.tasks = await this.fetchTasks()
-  }
+  }  
 }
 </script>
