+++ src/components/Footer.vue
@@ -1,7 +1,7 @@
 <template>
     <footer>
         <p>Copyright &copy; 2021</p>
-        <a href="/about">About</a>
+        <router-link to="/about">About</router-link>
     </footer>
 </template>
