+++ src/router/index.js
@@ -1,7 +1,13 @@
 import {createRouter, createWebHistory} from 'vue-router'
+import Home from '../views/Home'
 import About from '../views/About'
 
 const routes = [
+    {
+        path: '/',
+        name: 'Home',
+        component: Home
+    },
     {
         path: '/about',
         name: 'About',
@@ -13,3 +19,5 @@ const router = createRouter({
     history: createWebHistory(process.env.BASE_URL),
     routes
 })
+
+export default router
