+++ src/views/About.vue
@@ -1,3 +1,4 @@
 <template>
-    <h1>About</h1>
+    <h3>Version 1.0.0</h3>
+    <router-link to="./">Go Back</router-link>
 </template>
